+++ Pastebin/Database/DefaultRepositories/AmazonS3TextRepository.cs
@@ -4,14 +4,16 @@ using Pastebin.Interfaces;
 
 namespace Pastebin.Database.DefaultRepositories;
 
-public class AmazonS3TextRepository(IAmazonS3 amazonS3Client, IConfiguration configuration) : ITextRepository
+public class AmazonS3TextRepository(
+    IAmazonS3 amazonS3Client,
+    IConfiguration configuration) : ITextRepository
 {
     public async Task PostAsync(string key, string text)
     {
         var putObjectRequest = new PutObjectRequest
         {
             Key = key,
-            BucketName = configuration.GetSection("BucketName").Value,
+            BucketName = configuration.GetValue<string>("BucketName"),
             ContentBody = text
         };
         putObjectRequest.Metadata.Add("Content-Type", "text/plain");
@@ -23,9 +25,12 @@ public class AmazonS3TextRepository(IAmazonS3 amazonS3Client, IConfiguration con
         var getObjectRequest = new GetObjectRequest
         {
             Key = key,
-            BucketName = configuration.GetSection("BucketName").Value
+            BucketName = configuration.GetValue<string>("BucketName")
         };
+
         var s3Object = await amazonS3Client.GetObjectAsync(getObjectRequest);
+        if (s3Object is null)
+            return null;
 
         using var sr = new StreamReader(s3Object.ResponseStream);
         var text = await sr.ReadToEndAsync();
@@ -38,14 +43,14 @@ public class AmazonS3TextRepository(IAmazonS3 amazonS3Client, IConfiguration con
         var deleteObjectRequest = new DeleteObjectRequest
         {
             Key = key,
-            BucketName = configuration.GetSection("BucketName").Value
+            BucketName = configuration.GetValue<string>("BucketName")
         };
         await amazonS3Client.DeleteObjectAsync(deleteObjectRequest);
 
         var putObjectRequest = new PutObjectRequest
         {
             Key = key,
-            BucketName = configuration.GetSection("BucketName").Value,
+            BucketName = configuration.GetValue<string>("BucketName"),
             ContentBody = text
         };
         putObjectRequest.Metadata.Add("Content-Type", "text/plain");
@@ -57,7 +62,7 @@ public class AmazonS3TextRepository(IAmazonS3 amazonS3Client, IConfiguration con
         var deleteObjectRequest = new DeleteObjectRequest
         {
             Key = key,
-            BucketName = configuration.GetSection("BucketName").Value
+            BucketName = configuration.GetValue<string>("BucketName")
         };
         await amazonS3Client.DeleteObjectAsync(deleteObjectRequest);
     }
@@ -65,10 +70,12 @@ public class AmazonS3TextRepository(IAmazonS3 amazonS3Client, IConfiguration con
     public async Task DeleteMultipleByKeysListAsync(List<string> keys)
     {
         var deleteObjectsRequest = new DeleteObjectsRequest
-            { BucketName = configuration.GetSection("BucketName").Value };
+            { BucketName = configuration.GetValue<string>("BucketName") };
 
-        foreach (var key in keys) deleteObjectsRequest.AddKey(key);
+        foreach (var key in keys)
+            deleteObjectsRequest.AddKey(key);
 
-        await amazonS3Client.DeleteObjectsAsync(deleteObjectsRequest);
+        if (deleteObjectsRequest.Objects.Count > 0)
+            await amazonS3Client.DeleteObjectsAsync(deleteObjectsRequest);
     }
 }
